Rename DialogueMonoSystem to StoryMonoSystem
