+++ Assets/Prefabs/MonoSystems/Story/StoryMonoSystem.prefab.meta
