+++ Assets/Scenes/Test/StoryMonoSystem_Test.meta
