+++ Assets/Scenes/Test/StoryMonoSystem_Test/ClickMe_Test.mat.meta
