+++ Assets/Scripts/Runtime/AkashicGameManager.cs
@@ -1,6 +1,5 @@
 using Akashic.Core;
 using Akashic.Runtime.MonoSystems.BattleStates;
-using Akashic.Runtime.MonoSystems.Dialogue;
 using Akashic.Runtime.MonoSystems.ExplorationStates;
 using Akashic.Runtime.MonoSystems.GameStates;
 using Akashic.Runtime.MonoSystems.Party;
@@ -8,6 +7,7 @@ using Akashic.Runtime.MonoSystems.PlayerPrefs;
 using Akashic.Runtime.MonoSystems.Save;
 using Akashic.Runtime.MonoSystems.Scene;
 using Akashic.Runtime.MonoSystems.Sound;
+using Akashic.Runtime.MonoSystems.Story;
 using UnityEngine;
 
 namespace Akashic.Runtime
@@ -28,7 +28,7 @@ namespace Akashic.Runtime
         [SerializeField] private PartyMonoSystem partyMonoSystem;
         [SerializeField] private PlayerPreferencesMonoSystem playerPreferencesMonoSystem;
         [SerializeField] private SaveMonoSystem saveMonoSystem;
-        [SerializeField] private DialogueMonoSystem dialogueMonoSystem;
+        [SerializeField] private StoryMonoSystem storyMonoSystem;
 
         protected override string GetApplicationName()
         {
@@ -53,7 +53,7 @@ namespace Akashic.Runtime
             AddMonoSystem<PartyMonoSystem, IPartyMonoSystem>(partyMonoSystem);
             AddMonoSystem<PlayerPreferencesMonoSystem, IPlayerPreferencesMonoSystem>(playerPreferencesMonoSystem);
             AddMonoSystem<SaveMonoSystem, ISaveMonoSystem>(saveMonoSystem);
-            AddMonoSystem<DialogueMonoSystem, IDialogueMonoSystem>(dialogueMonoSystem);
+            AddMonoSystem<StoryMonoSystem, IStoryMonoSystem>(storyMonoSystem);
         }
     }
 }
