+++ Assets/Scripts/Runtime/MonoSystems/Story/Messages.cs
@@ -1,7 +1,7 @@
 using Akashic.Core;
 using Akashic.ScriptableObjects.Scripts;
 
-namespace Akashic.Runtime.MonoSystems.Dialogue
+namespace Akashic.Runtime.MonoSystems.Story
 {
     internal sealed class DialogueStoryEventMessage : IMessage
     {
