+++ Assets/Scripts/Runtime/MonoSystems/Story/StoryPoint.cs
@@ -1,7 +1,7 @@
 using System;
 using UnityEngine;
 
-namespace Akashic.Runtime.MonoSystems.Dialogue
+namespace Akashic.Runtime.MonoSystems.Story
 {
     [Serializable]
     internal sealed class StoryPoint
